fixed bugs! It works yay

+++ project/sec.h
@@ -16,7 +16,7 @@ void output_sec(uint8_t* buf, size_t length);
 uint32_t TLV_maker(uint8_t* buf, uint8_t type, uint16_t length, uint8_t* value);
 int find_location(uint8_t* data, int state_sec, uint8_t type);
 uint16_t get_size(uint8_t* data, int bufloc);
-uint16_t plaintxt_length(uint16_t payload_length);
+int plaintxt_length(uint16_t payload_length);
 typedef struct // maybe useful to cast pointer into this
 {
     uint8_t type;
